# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gateway Apis terabyte",
        "version": "1.0.0"
    },
    "paths": {
        "/user": {},
        "/api/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Pegar todos usuarios",
                "operationId": "ff9c7b24f765e92a815d3a4f5efbf114",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Adicionar usuario",
                "operationId": "359bc743f9c9e92c8493e911d8dc3dd3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Matheus Sodré",
                                    "email": "matheus@email.com",
                                    "password": "senha12324"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ideentify": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "permissions": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "The email has already been taken"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/{uuid}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Pegar usuario pelo uuid",
                "description": "Returns project data",
                "operationId": "c6a6cf290ca4eb8dfc0a51f0ee7e7c39",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Usuario uuid",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "valor UUID",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Atualizar usuario",
                "operationId": "c0207a41139f2df43dd5b54fcfb5b476",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "valor UUID",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Deletar usuario",
                "description": "Deletar usuario",
                "operationId": "ad26c8aca2656684f3b029dd7c512f01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of pet to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "pet deleted"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}